test_api_solution/test_case/requirements.txt
flake8==5.0.4
flake8-docstrings==1.7.0
Django==3.2.16
djangorestframework==3.12.4
djoser==2.2.2
python-dotenv==1.0.1
requests 
beautifulsoup4